+++ circleseq/mergeReads.py
@@ -2,20 +2,28 @@ from __future__ import print_function
 import argparse
 import itertools
 import gzip
+import logging
+import os
 from utility import reverseComplement, fq
+
+
 
 def mergeReads(fastq1_filename, fastq2_filename, out):
     fastq1_file = fq(fastq1_filename)
     fastq2_file = fq(fastq2_filename)
-
-    with gzip.open(out, 'wb') as o:
-        for r1, r2 in itertools.izip(fastq1_file, fastq2_file):
-            merged_sequence = reverseComplement(r1[1]) + r2[1]
-            merged_quality_scores = r1[3][::-1] + r2[3]
-            print(r1[0], file=o)
-            print(merged_sequence, file=o)
-            print(r1[2], file=o)
-            print(merged_quality_scores, file=o)
+    if not os.path.exists(out):
+        print("Merged file does not exist. Merging...", file=sys.stderr)
+        with gzip.open(out, 'wb') as o:
+            for r1, r2 in itertools.izip(fastq1_file, fastq2_file):
+                merged_sequence = reverseComplement(r1[1]) + r2[1]
+                merged_quality_scores = r1[3][::-1] + r2[3]
+                print(r1[0], file=o)
+                print(merged_sequence, file=o)
+                print(r1[2], file=o)
+                print(merged_quality_scores, file=o)
+    else:
+        print("Merged file exists.")
+        pass
 
 def main():
     parser = argparse.ArgumentParser(description='Merge CIRCLE-seq reads for alignment.')
